% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{pythonic module import function}
\usage{
import(path)
}
\description{
This function allows you to import a R file as a module like in python
so that its content won't contaminate your working directory.
}
